move swigluoai_and_mul kernel to activation.cpp
